+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Deprecated
 
 ### Removed
+- Drop empty constructors (#320)
 - Drop 7.6-specific code (#305, #314)
 - Drop unneeded Travis CI configuration settings (#295, #296, #297, #298)
 - Remove deprecated methods from `TestingFramework` (#285)